// src/app/edit_user/types.ts
export interface UserFormData {
    firstname: string;
    middlename?: string;
    lastname: string;
    honorificprefix?: string;
    honorificsuffix?: string;
    jobtitle: string;
    company: string;
    logo: string;
    website?: string;
    cellphone?: string;
    whatsapp?: string;
    viber?: string;
    wechat?: string;
    workphone?: string;
    workemail: string;
    workfax?: string;
    address: string;
    location: string;
    linkedin?: string;
    facebook?: string;
    note?: string;
    profileImage?: string;
    street: string;
    city: string;
    state: string;
    postalCode: string;
    country: string;
    factoryLocation?: string;
    factoryStreet?: string;
    factoryCity?: string;
    factoryState?: string;
    factoryPostalCode?: string;
    factoryCountry?: string;
  }
  